REFACTOR: Moved messages and tests of staking module to own files.
